pyproject.toml
[project]
name = "shallwego"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.5.2",
    "langchain>=0.3.25",
    "langchain-core>=0.3.60",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "pylint>=3.3.7",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "streamlit-cookies-controller>=0.0.4",
]
